# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AlertContacts API",
        "description": "API de sécurité personnelle pour protéger et rassurer ses proches. Cette API permet de gérer les zones de sécurité, zones de danger, invitations, relations entre utilisateurs et notifications en temps réel.",
        "contact": {
            "name": "Support AlertContacts",
            "email": "support@alertcontacts.net"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://alertcontacts.net/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://mobile.alertcontacts.net/api",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/safe-zones": {
            "get": {
                "tags": [
                    "Safe Zones"
                ],
                "summary": "Récupérer les zones de sécurité",
                "description": "Récupère toutes les zones de sécurité actives de l'utilisateur connecté",
                "operationId": "02a8af6911eb73304d6027c86e0e098d",
                "responses": {
                    "200": {
                        "description": "Liste des zones de sécurité récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SafeZone"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Safe Zones"
                ],
                "summary": "Créer une zone de sécurité",
                "description": "Crée une nouvelle zone de sécurité (cercle ou polygone) avec assignation optionnelle de contacts",
                "operationId": "23ee04fd1d24e8e39b70a5b8220cc3e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom de la zone",
                                        "type": "string",
                                        "example": "Maison"
                                    },
                                    "icon": {
                                        "description": "Icône de la zone",
                                        "type": "string",
                                        "example": "home"
                                    },
                                    "center": {
                                        "description": "Centre de la zone (pour les cercles)",
                                        "properties": {
                                            "lat": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 48.8566
                                            },
                                            "lng": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 2.3522
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "radius_m": {
                                        "description": "Rayon en mètres (pour les cercles)",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "geom": {
                                        "description": "Géométrie de la zone (pour les polygones)",
                                        "properties": {
                                            "coordinates": {
                                                "type": "array",
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "number",
                                                            "format": "float"
                                                        }
                                                    }
                                                }
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "contact_ids": {
                                        "description": "IDs des contacts à assigner à la zone",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "active_hours": {
                                        "description": "Heures d'activité de la zone",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Zone de sécurité créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SafeZone"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SafeZone": {
                "title": "Zone de Sécurité",
                "description": "Zone privée de sécurité créée par un utilisateur",
                "properties": {
                    "id": {
                        "description": "Identifiant unique de la zone",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de la zone de sécurité",
                        "type": "string",
                        "example": "Maison"
                    },
                    "description": {
                        "description": "Description de la zone",
                        "type": "string",
                        "example": "Zone autour de la maison familiale",
                        "nullable": true
                    },
                    "center": {
                        "description": "Coordonnées du centre de la zone",
                        "properties": {
                            "lat": {
                                "description": "Latitude",
                                "type": "number",
                                "format": "float",
                                "example": 48.8566
                            },
                            "lng": {
                                "description": "Longitude",
                                "type": "number",
                                "format": "float",
                                "example": 2.3522
                            }
                        },
                        "type": "object"
                    },
                    "radius_meters": {
                        "description": "Rayon de la zone en mètres",
                        "type": "integer",
                        "example": 100
                    },
                    "icon_key": {
                        "description": "Clé de l'icône à afficher",
                        "type": "string",
                        "example": "home"
                    },
                    "address": {
                        "description": "Adresse de la zone",
                        "type": "string",
                        "example": "123 Rue de la Paix, Paris",
                        "nullable": true
                    },
                    "member_ids": {
                        "description": "IDs des membres assignés à cette zone",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            2,
                            3
                        ]
                    },
                    "owner_id": {
                        "description": "ID du propriétaire de la zone",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum. Utilisez le token obtenu lors de la connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}